The Great Splittening begins!
The Actor class is refactored to use custom data models

We now have specific sheets for NPCs and PCs.

NPCs can use spells now, with no restrictions.
