+++ src/templates/actor/parts/actor-spells.hbs
@@ -1,4 +1,4 @@
-{{#each data.spellcasting as |class classKey|}}
+{{#each actor.system.spellcasting as |class classKey|}}
   <h2 class="features__title">
     <span>{{class.name}}</span>
     <span>
