+++ src/templates/actor/parts/pc-header.hbs
@@ -20,15 +20,15 @@
         <input
           type="number"
           name="system.hp.value"
-          value="{{data.hp.value}}"
-          max="{{data.hp.max}}"
+          value="{{actor.system.hp.value}}"
+          max="{{hp.max}}"
           data-dtype="Number"/>
         <span>/</span>
         <input
           type="number"
           name="system.hp.max"
-          min="{{data.hp.value}}"
-          value="{{data.hp.max}}"
+          min="{{actor.system.hp.value}}"
+          value="{{actor.system.hp.max}}"
           data-dtype="Number" />
       </div>
 
@@ -45,7 +45,7 @@
         <input
           type="number"
           name="system.ac"
-          value="{{ac}}"
+          value="{{actor.system.ac}}"
           readonly
           data-dtype="Number"/>
       </div>
@@ -63,7 +63,7 @@
         <input
           type="number"
           name="system.move"
-          value="{{move}}"
+          value="{{actor.system.move}}"
           readonly
           data-dtype="Number"/>
       </div>
